refactor(idea_service): extract dashboards (#163)
  before: 1579
  after:  1395 (-184)
